rename: rename constants
