--- src/constants/game.ts
@@ -1,9 +1,0 @@
-export const MAX_ROUND = 4;
-
-export const POSITION: Record<string, string> = {
-  frontend: '프론트엔드',
-  backend: '백엔드',
-  ios: 'iOS',
-  android: '안드로이드',
-  design: '디자인',
-};
